+++ src/backend/src/main/java/RunningMachines/R2R/domain/course/dto/CourseResponseDto.java
@@ -1,4 +1,4 @@
-package RunningMachines.R2R.domain.running.dto;
+package RunningMachines.R2R.domain.course.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,7 +7,9 @@ import java.util.List;
 
 @Getter
 @AllArgsConstructor
-public class GpxResponseDto {
+public class CourseResponseDto {
     private String fileName;
     private List<String> waypoints;
+    private double distance;
+    private List<String> tags;
 }
