+++ src/backend/src/main/java/RunningMachines/R2R/domain/course/entity/ReviewTag.java
@@ -1,0 +1,5 @@
+package RunningMachines.R2R.domain.course.entity;
+
+public enum ReviewTag {
+    Tag
+}
