+++ src/backend/src/main/java/RunningMachines/R2R/domain/course/repository/CourseRepository.java
@@ -1,6 +1,6 @@
-package RunningMachines.R2R.domain.running.repository;
+package RunningMachines.R2R.domain.course.repository;
 
-import RunningMachines.R2R.domain.running.entity.Course;
+import RunningMachines.R2R.domain.course.entity.Course;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CourseRepository extends JpaRepository<Course, Long> {
