--- src/backend/src/main/java/RunningMachines/R2R/domain/running/entity/ReviewTag.java
@@ -1,5 +1,0 @@
-package RunningMachines.R2R.domain.running.entity;
-
-public enum ReviewTag {
-    Tag
-}
